+++ test/e2e/rule_test.go
@@ -29,31 +29,21 @@ groups:
 `
 
 var (
-	ruleGossipSuite = newSpinupSuite().
-			Add(querier(1, ""), queryCluster(1)).
-			Add(ruler(1, alwaysFireRule)).
-			Add(ruler(2, alwaysFireRule)).
-			Add(alertManager(1), "")
-
 	ruleStaticFlagsSuite = newSpinupSuite().
-				Add(querierWithStoreFlags(1, "", rulerGRPC(1), rulerGRPC(2)), "").
+				Add(querierWithStoreFlags(1, "", rulerGRPC(1), rulerGRPC(2))).
 				Add(rulerWithQueryFlags(1, alwaysFireRule, queryHTTP(1))).
 				Add(rulerWithQueryFlags(2, alwaysFireRule, queryHTTP(1))).
-				Add(alertManager(1), "")
+				Add(alertManager(1))
 
 	ruleFileSDSuite = newSpinupSuite().
-			Add(querierWithFileSD(1, "", rulerGRPC(1), rulerGRPC(2)), "").
+			Add(querierWithFileSD(1, "", rulerGRPC(1), rulerGRPC(2))).
 			Add(rulerWithFileSD(1, alwaysFireRule, queryHTTP(1))).
 			Add(rulerWithFileSD(2, alwaysFireRule, queryHTTP(1))).
-			Add(alertManager(1), "")
+			Add(alertManager(1))
 )
 
 func TestRule(t *testing.T) {
 	for _, tt := range []testConfig{
-		{
-			"gossip",
-			ruleGossipSuite,
-		},
 		{
 			"staticFlag",
 			ruleStaticFlagsSuite,